chatbot_backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
nltk==3.9.1
rapidfuzz==3.9.6
python-dotenv==1.0.1
django-cors-headers==4.4.0
